move svg code
